requirements.txt
alembic==1.6.5
aniso8601==7.0.0
asgiref==3.4.1
asyncio==3.4.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
FastAPI-SQLAlchemy==0.2.1
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
Mako==1.1.5
MarkupSafe==2.0.1
promise==2.3
psycopg2==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
